add asset proxy file upload script
