+++ package.json
@@ -9,6 +9,7 @@
     "lint": "prettier src/ test/ --write"
   },
   "dependencies": {
+    "@dfinity/assets": "^1.3.0",
     "@dfinity/candid": "^1.0.1",
     "@dfinity/principal": "^1.0.1",
     "azle": "^0.20.2"
