+++ scripts/deploy.ts
@@ -8,65 +8,13 @@ import { createActor as createProvisionActor } from "../test/dfx_generated/provi
 import { Principal } from "@dfinity/principal";
 import { configureCanisters } from "../test/src/utils/deploy";
 import { managementIdl, managementService } from "../test/src/utils/canister";
-
-async function getAgent(): Promise<HttpAgent> {
-  const host = process.env.DFX_NETWORK === 'ic' ? 'https://icp-api.io' : 'http://127.0.0.1:4943';
-  const identityPemFilePath = process.env.IDENTITY_PEM_FILE!;
-  const identity = Secp256k1KeyIdentity.fromPem(
-    readFileSync(identityPemFilePath, {encoding: 'utf8'})
-  )
-  
-  const agent = new HttpAgent({ host, identity: identity as Identity });
-  if ( process.env.DFX_NETWORK !== 'ic' ) {
-    await agent.fetchRootKey().catch((err) => {
-      console.warn(
-        "Unable to fetch root key. Check to ensure that your local replica is running"
-      );
-      console.error(err);
-    });
-  }
-
-  return agent;
-}
-
-function exec(cmd: string): Promise<string> {
-  return new Promise((resolve, reject) => {
-    execCommand(cmd, {
-      encoding: 'utf8',
-    }, (err, stdout, stderr) => {
-      if ( err ) reject(stderr);
-      resolve(stdout.trim());
-    });
-  });
-}
-
-async function deployCanister(canisterName: string) {
-  console.log(`Deploying ${canisterName}...`);
-  const networkFlag = process.env.DFX_NETWORK === 'ic' ? '--ic' : '';
-  await exec(`dfx deploy ${canisterName} ${networkFlag}`);
-  const id = await exec(`dfx canister id ${canisterName} ${networkFlag}`);
-  console.log(`Deployed ${canisterName} canister: ${id}`);
-
-  return id;
-}
-
-async function buildCanister(canisterName: string) {
-  console.log(`Building ${canisterName}...`);
-  await exec(`dfx build --check ${canisterName}`);
-  console.log(`Build success for ${canisterName} canister`);
-}
-
-async function getCanisterId(canisterName: string) {
-  const networkFlag = process.env.DFX_NETWORK === 'ic' ? '--ic' : '';
-  const id = await exec(`dfx canister id ${canisterName} ${networkFlag}`);
-  return id;
-}
+import { buildCanister, getAgent, getCanisterId } from "./common";
 
 async function main() {
   const agent = await getAgent();
-  const provisionCanisterId = Principal.fromText(await deployCanister('provision'));
-  const assetProxyCanisterId = Principal.fromText(await deployCanister('asset_proxy'));
-  const tempAssetCanisterId = Principal.fromText(await deployCanister('asset'));
+  const provisionCanisterId = Principal.fromText(await getCanisterId('provision'));
+  const assetProxyCanisterId = Principal.fromText(await getCanisterId('asset_proxy'));
+  const tempAssetCanisterId = Principal.fromText(await getCanisterId('asset'));
   const managementCanisterId = Principal.fromText('aaaaa-aa');
   await buildCanister("estate_dao_nft");
