+++ scripts/file-upload-asset-proxy.ts
@@ -1,0 +1,23 @@
+import { readFile } from "fs/promises";
+import { AssetManager } from "@dfinity/assets";
+import { getAgent, getCanisterId } from "./common";
+
+async function main() {
+  const assetProxyManager = new AssetManager({
+    canisterId: await getCanisterId('asset_proxy'),
+    agent: await getAgent(),
+  });
+ 
+  const assetManager = new AssetManager({
+    canisterId: await getCanisterId('asset'),
+    agent: await getAgent(),
+  });
+ 
+  const file = await readFile('./test.jpg');
+  console.log(await assetProxyManager.store('./test.jpg', {
+    contentEncoding: 'identity'
+  }));
+  console.log(await assetManager.get('/test.jpg', ['identity']))
+}
+
+main();
